+++ ADMF/functions/Invoke-AdmfForest.ps1
@@ -41,7 +41,7 @@
 		Applies the full forest configuration to the contoso.com domain.
 	#>
 	[Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingPlainTextForPassword', '')]
-	[CmdletBinding(SupportsShouldProcess = $true)]
+	[CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'High')]
 	Param (
 		[PSFComputer]
 		$Server,
@@ -60,7 +60,7 @@
 	{
 		$parameters = $PSBoundParameters | ConvertTo-PSFHashtable -Include Server, Credential
 		$originalArgument = Invoke-PreCredentialProvider @parameters -ProviderName $CredentialProvider -Parameter $parameters -Cmdlet $PSCmdlet
-		try { $dcServer = Resolve-DomainController @parameters }
+		try { $dcServer = Resolve-DomainController @parameters -Confirm:$false }
 		catch {
 			Invoke-PostCredentialProvider -ProviderName $CredentialProvider -Server $originalArgument.Server -Credential $originalArgument.Credential -Cmdlet $PSCmdlet
 			throw