+++ library/ADMF/ADMF/UpdateDCOptions.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ADMF
+{
+    [Flags]
+    public enum UpdateDCOptions
+    {
+        Share = 1,
+        FSAccessRule = 2,
+
+        Default = Share | FSAccessRule,
+        All = Share | FSAccessRule
+    }
+}