+++ ADMF/changelog.md
@@ -1,5 +1,17 @@
 ﻿# Changelog
 
+## 1.2.36 (2020-07-31)
+
+- New: Invoke-AdmfDC - new command to apply DC configuration entries on DCs
+- New: Test-AdmfDC - new command to test DC configuration entries
+- Upd: Invoke-AdmfDomain - added confirm-impact high
+- Upd: Invoke-AdmfForest - added confirm-impact high
+- Upd: Get-AdmfContext - added -DomainTable switch parameter to get a list of which server/domain has what Contexts assigned.
+- Upd: Set-AdmfContext - storing Context selection per domain, not per server.
+- Upd: Configuration - added setting 'ADMF.DCInstall.Context.Prompt.Enable' to allow disabling Context prompts for Install commands in DCManagement
+- Upd: Set-AdmfContext - add support for file system accesss rules Component
+- Fix: Invoke-Admf*, Test-Admf* - will no longer prompt for confirmation when resolving domain controller to use.
+
 ## 1.1.27 (2020-06-21)
 
 - Upd: Set-AdmfContext - added Component: DC > Shares